Organize test reporters into their own directory
