+++ sorbet/rbi/shims/test_reporter.rbi
@@ -1,6 +1,6 @@
 # typed: true
 
-class RubyLsp::TestReporter
+class RubyLsp::LspReporter
   # https://code.visualstudio.com/api/references/vscode-api#Position
   Position = T.type_alias { { line: Integer, character: Integer } }
 
@@ -12,4 +12,17 @@ class RubyLsp::TestReporter
 
   # https://code.visualstudio.com/api/references/vscode-api#StatementCoverage
   StatementCoverage = T.type_alias { { executed: Integer, location: Position, branches: T::Array[BranchCoverage] } }
+end
+
+class ::Test::Unit::UI::Console::TestRunner
+  def initialize(suite, options = T.unsafe(nil))
+    @mediator = T.let(T.unsafe(nil), Test::Unit::UI::TestRunnerMediator)
+  end
+end
+
+class RubyLsp::TestUnitReporter
+  def initialize(suite, options = T.unsafe(nil))
+    @current_uri = T.let(nil, T.nilable(URI::Generic))
+    @current_test_id = T.let(nil, T.nilable(String))
+  end
 end
